Added description and gazebo packages
