--- rae_bringup/config/cal.json
@@ -1,198 +1,0 @@
-{
-    "batchName": "",
-    "batchTime": 0,
-    "boardConf": "",
-    "boardCustom": "",
-    "boardName": "RAE",
-    "boardOptions": 0,
-    "boardRev": "R1M0E1",
-    "cameraData": [
-        [
-            1,
-            {
-                "cameraType": 0,
-                "distortionCoeff": [
-                    2.8254761695861816,
-                    7.1769328117370605,
-                    3.336625741212629e-05,
-                    0.00010876730812015012,
-                    0.9267292022705078,
-                    3.173186779022217,
-                    8.097966194152832,
-                    2.9521195888519287,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0
-                ],
-                "extrinsics": {
-                    "rotationMatrix": [
-                        [
-                            0.9975343346595764,
-                            -0.05187312141060829,
-                            -0.0472693145275116
-                        ],
-                        [
-                            0.052037063986063004,
-                            0.99864262342453,
-                            0.0022435400169342756
-                        ],
-                        [
-                            0.04708877205848694,
-                            -0.004697764758020639,
-                            0.9988796710968018
-                        ]
-                    ],
-                    "specTranslation": {
-                        "x": -7.5,
-                        "y": 0.0,
-                        "z": 0.0
-                    },
-                    "toCameraSocket": 2,
-                    "translation": {
-                        "x": -7.456183433532715,
-                        "y": -0.24615898728370667,
-                        "z": -0.4226444363594055
-                    }
-                },
-                "height": 800,
-                "intrinsicMatrix": [
-                    [
-                        575.981201171875,
-                        0.0,
-                        639.8970336914063
-                    ],
-                    [
-                        0.0,
-                        575.95458984375,
-                        395.1063537597656
-                    ],
-                    [
-                        0.0,
-                        0.0,
-                        1.0
-                    ]
-                ],
-                "lensPosition": 0,
-                "specHfovDeg": 110.0,
-                "width": 1280
-            }
-        ],
-        [
-            2,
-            {
-                "cameraType": 0,
-                "distortionCoeff": [
-                    4.2559967041015625,
-                    2.996381998062134,
-                    0.0005118183325976133,
-                    -0.000867578899487853,
-                    0.18907761573791504,
-                    4.593552589416504,
-                    4.3449931144714355,
-                    0.8202019333839417,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0
-                ],
-                "extrinsics": {
-                    "rotationMatrix": [],
-                    "specTranslation": {
-                        "x": 0.0,
-                        "y": 0.0,
-                        "z": 0.0
-                    },
-                    "toCameraSocket": -1,
-                    "translation": {
-                        "x": 0.0,
-                        "y": 0.0,
-                        "z": 0.0
-                    }
-                },
-                "height": 800,
-                "intrinsicMatrix": [
-                    [
-                        570.9515380859375,
-                        0.0,
-                        623.2518920898438
-                    ],
-                    [
-                        0.0,
-                        570.8521728515625,
-                        396.2295837402344
-                    ],
-                    [
-                        0.0,
-                        0.0,
-                        1.0
-                    ]
-                ],
-                "lensPosition": 0,
-                "specHfovDeg": 110.0,
-                "width": 1280
-            }
-        ]
-    ],
-    "hardwareConf": "",
-    "imuExtrinsics": {
-        "rotationMatrix": [],
-        "specTranslation": {
-            "x": 0.0,
-            "y": 0.0,
-            "z": 0.0
-        },
-        "toCameraSocket": -1,
-        "translation": {
-            "x": 0.0,
-            "y": 0.0,
-            "z": 0.0
-        }
-    },
-    "miscellaneousData": [],
-    "productName": "",
-    "stereoRectificationData": {
-        "leftCameraSocket": 2,
-        "rectifiedRotationLeft": [
-            [
-                0.9978554248809814,
-                0.032943274825811386,
-                0.056562189012765884
-            ],
-            [
-                -0.032864656299352646,
-                0.9994571208953857,
-                -0.0023198111448436975
-            ],
-            [
-                -0.05660790205001831,
-                0.00045593924005515873,
-                0.9983963966369629
-            ]
-        ],
-        "rectifiedRotationRight": [
-            [
-                0.9997727870941162,
-                -0.019129034131765366,
-                0.009404786862432957
-            ],
-            [
-                0.019115954637527466,
-                0.9998161792755127,
-                0.001478599733673036
-            ],
-            [
-                -0.009431342594325542,
-                -0.001298482296988368,
-                0.9999547004699707
-            ]
-        ],
-        "rightCameraSocket": 1
-    },
-    "version": 7
-}
-
