--- rae_bringup/launch/rae_camera.launch.py
@@ -1,48 +1,0 @@
-import os
-
-from ament_index_python.packages import get_package_share_directory
-from launch import LaunchDescription
-from launch.actions import DeclareLaunchArgument, OpaqueFunction
-from launch.substitutions import LaunchConfiguration
-from launch_ros.actions import ComposableNodeContainer
-from launch_ros.descriptions import ComposableNode
-
-def launch_setup(context, *args, **kwargs):
-    log_level = 'info'
-    if(context.environment.get('DEPTHAI_DEBUG')=='1'):
-        log_level='debug'
-    params_file = LaunchConfiguration("params_file")
-    depthai_prefix = get_package_share_directory("depthai_ros_driver")
-
-    name = LaunchConfiguration('name').perform(context)
-
-    return [
-         ComposableNodeContainer(
-            name=name+"_container",
-            namespace="",
-            package="rclcpp_components",
-            executable="component_container",
-            composable_node_descriptions=[
-                    ComposableNode(
-                        package="depthai_ros_driver",
-                        plugin="depthai_ros_driver::Camera",
-                        name=name,
-                        parameters=[params_file],
-                    )
-            ],
-            arguments=['--ros-args', '--log-level', log_level],
-            output="both",
-        ),
-    ]
-
-
-def generate_launch_description():
-    rae_prefix = get_package_share_directory("rae_bringup")
-    declared_arguments = [
-        DeclareLaunchArgument("name", default_value="rae"),
-        DeclareLaunchArgument("params_file", default_value=os.path.join(rae_prefix, 'config', 'camera.yaml')),
-    ]
-
-    return LaunchDescription(
-        declared_arguments + [OpaqueFunction(function=launch_setup)]
-    )
