+++ rae_description/launch/rsp.launch.py
@@ -1,0 +1,43 @@
+import os
+
+from ament_index_python.packages import get_package_share_directory
+
+from launch import LaunchDescription
+from launch.actions import DeclareLaunchArgument
+from launch.substitutions import Command, LaunchConfiguration
+from launch_ros.actions import Node
+
+def generate_launch_description():
+
+    use_sim_time = LaunchConfiguration('use_sim_time') 
+    pkg_path = os.path.join(get_package_share_directory('rae_description'))
+    xacro_path = os.path.join(pkg_path, 'urdf', 'rae.urdf.xacro')
+
+    rsp_node = Node(
+        package='robot_state_publisher',
+        executable='robot_state_publisher',
+        name='robot_state_publisher',
+        parameters=[{
+            'robot_description': Command(['xacro ', xacro_path, ' sim_mode:=', use_sim_time]),
+            'use_sim_time': use_sim_time
+        }]
+    )
+
+    jsp_node = Node(
+        package='joint_state_publisher',
+        executable='joint_state_publisher',
+        name='joint_state_publisher',
+        parameters=[{
+            'use_sim_time': use_sim_time
+        }]
+    )
+
+    return LaunchDescription([
+        DeclareLaunchArgument(
+            'use_sim_time',
+            default_value='false',
+            description='Use sim time'
+        ),
+        rsp_node,
+        jsp_node
+    ])
