+++ rae_description/urdf/camera.urdf.xacro
@@ -1,0 +1,57 @@
+<?xml version="1.0"?>
+<robot xmlns:xacro="http://www.ros.org/wiki/xacro" name="robot">
+
+    <plugin filename="ignition-gazebo-sensors-system" name="ignition::gazebo::systems::Sensors">
+        <render_engine>ogre2</render_engine>
+    </plugin>
+
+    <link name="rgb_camera_link">
+        <xacro:body mass="0.001" d="0.010" h="0.03" w="0.03">
+            <origin xyz="0.0 0.0 0.0" rpy="0.0 0.0 0.0" />
+        </xacro:body>
+        <collision>
+            <geometry>
+                <box size="0.010 0.03 0.03"/>
+            </geometry>
+        </collision>
+    </link>
+
+    <joint name="rgb_camera_joint" type="fixed">
+        <parent link="base_link" />
+        <child link="rgb_camera_link" />
+        <origin xyz="0.09 0.0 0.025" rpy="0 0 0" />
+    </joint>
+
+    <link name="rgb_camera_link_optical_frame">
+        <xacro:body mass="0.00001" d="0.0000010" h="0.000003" w="0.00003">
+            <origin xyz="0.0 0.0 0.0" rpy="0.0 0.0 0.0" />
+        </xacro:body>
+    </link>
+
+    <joint name="rgb_camera_link_optical_frame_joint" type="fixed">
+        <origin xyz="0.0051 0 0" rpy="0.0 0.0 0.0" />
+        <parent link="rgb_camera_link" />
+        <child link="rgb_camera_link_optical_frame" />
+    </joint>
+
+    <gazebo reference="rgb_camera_link_optical_frame">
+        <sensor name="rgbd_camera" type="camera">
+            <camera>
+                <horizontal_fov>1.047</horizontal_fov>
+                <image>
+                    <width>320</width>
+                    <height>240</height>
+                </image>
+                <clip>
+                    <near>0.1</near>
+                    <far>100</far>
+                </clip>
+            </camera>
+            <ignition_frame_id>rgb_camera_link_optical_frame</ignition_frame_id>
+            <always_on>1</always_on>
+            <update_rate>30</update_rate>
+            <visualize>true</visualize>
+            <topic>/camera/image</topic>
+        </sensor>
+    </gazebo>
+</robot>
