+++ rae_description/urdf/inertial_moment_calc.xacro
@@ -5,23 +5,24 @@
         <inertial>
             <xacro:insert_block name="origin"/>
             <mass value="${mass}" />
-            <inertia ixx="${(1/12) * mass * (h*h+w*w)}" 
-                     ixy="0.0"
-                     ixz="0.0"
-                     iyy="${(1/12) * mass * (d*d+w*w)}" 
-                     iyz="0.0"
-                     izz="${(1/12) * mass * (d*d+h*h)}" />
+            <inertia 
+                ixx="${(1/12) * mass * (h*h+w*w)}" 
+                ixy="0.0"
+                ixz="0.0"
+                iyy="${(1/12) * mass * (d*d+w*w)}" 
+                iyz="0.0"
+                izz="${(1/12) * mass * (d*d+h*h)}" />
         </inertial>
     </xacro:macro>
-
 
     <xacro:macro name="wheels" params="mass h r *origin">
         <inertial>
             <xacro:insert_block name="origin"/>
             <mass value="${mass}" />
-            <inertia ixx="${(1/2) * mass * (r*r)}"  ixy="0.0" ixz="0.0"
-                    iyy="${(1/12) * mass * (3*r*r + h*h)}" iyz="0.0"
-                    izz="${(1/12) * mass * (3*r*r + h*h)}"/>
+            <inertia 
+                ixx="${(1/2) * mass * (r*r)}"  ixy="0.0" ixz="0.0"
+                iyy="${(1/12) * mass * (3*r*r + h*h)}" iyz="0.0"
+                izz="${(1/12) * mass * (3*r*r + h*h)}"/>
         </inertial>
     </xacro:macro>
 
@@ -29,11 +30,11 @@
         <inertial>
             <xacro:insert_block name="origin"/>
             <mass value="${mass}" />
-            <inertia ixx="${(2/5) * mass * (r*r)}"  ixy="0.0" ixz="0.0"
-                    iyy="${(2/5) * mass * (r*r)}" iyz="0.0"
-                    izz="${(2/5) * mass * (r*r)}"/>
+            <inertia 
+                ixx="${(2/5) * mass * (r*r)}"  ixy="0.0" ixz="0.0"
+                iyy="${(2/5) * mass * (r*r)}" iyz="0.0"
+                izz="${(2/5) * mass * (r*r)}"/>
         </inertial>
     </xacro:macro>
-
 
 </robot>
