--- rae_description/urdf/rae_control.xacro
@@ -1,42 +1,0 @@
-<?xml version="1.0"?>
-<robot name="rae"
-  xmlns:xacro="http://ros.org/wiki/xacro">
-  <ros2_control name="RAE" type="system">
-    <hardware>
-      <plugin>rae_hw/RaeHW</plugin>
-      <param name="left_wheel_name">left_wheel_joint</param>
-      <param name="right_wheel_name">right_wheel_joint</param>
-      <param name="pwmL">19</param>
-      <param name="pwmR">20</param>
-      <param name="phL">41</param>
-      <param name="phR">45</param>
-      <param name="enLA">42</param>
-      <param name="enLB">43</param>
-      <param name="enRA">46</param>
-      <param name="enRB">47</param>
-      <param name="encTicsPerRevL">187</param>
-      <param name="encTicsPerRevR">187</param>
-      <param name="maxVelL">32</param>
-      <param name="maxVelR">32</param>
-      <param name="loop_rate">30</param>
-      <param name="chip_name">gpiochip0</param>
-    </hardware>
-    <joint name="left_wheel_joint">
-      <command_interface name="velocity">
-        <param name="min">-1.68</param>
-        <param name="max">1.68</param>
-      </command_interface>
-      <state_interface name="velocity"/>
-      <state_interface name="position"/>
-    </joint>
-    <joint name="right_wheel_joint">
-      <command_interface name="velocity">
-        <param name="min">-1.68</param>
-        <param name="max">1.68</param>
-      </command_interface>
-      <state_interface name="velocity"/>
-      <state_interface name="position"/>
-    </joint>
-  </ros2_control>
-
-</robot>
