--- rae_description/urdf/rae_gazebo.xacro
@@ -1,92 +1,0 @@
-<?xml version="1.0"?>
-<robot name="rae"
-  xmlns:xacro="http://ros.org/wiki/xacro">
-  <xacro:include filename="inertial_moment_calc.xacro"/>
-  <xacro:property name="deg_to_rad" value="0.01745329251994329577" />
- 
-  <gazebo>
-    <plugin filename="libignition-gazebo-diff-drive-system.so" name="ignition::gazebo::systems::DiffDrive">
-      <frame_id>odom</frame_id>
-      <child_frame_id>base_footprint</child_frame_id>
-      <tf_topic>tf</tf_topic>
-      <publish_odom>true</publish_odom>
-      <publish_odom_tf>true</publish_odom_tf>
-      <publish_wheel_tf>true</publish_wheel_tf>
-      <odom_topic>odom</odom_topic>
-      <left_joint>left_wheel_joint</left_joint>
-      <right_joint>right_wheel_joint</right_joint>
-      <wheel_separation>0.0854</wheel_separation>
-      <!-- check correct wheel radius when you come back to the task -->
-      <wheel_radius>0.0119</wheel_radius>
-      <odom_publish_frequency>50</odom_publish_frequency>
-      <topic>cmd_vel</topic>
-    </plugin>
-
-  </gazebo>
-
-  <gazebo reference="rgb_camera_link_optical_frame">
-    <sensor name="rgbd_camera" type="rgbd_camera">
-      <camera>
-        <horizontal_fov>1.047</horizontal_fov>
-        <image>
-          <width>320</width>
-          <height>240</height>
-        </image>
-        <clip>
-          <near>0.1</near>
-          <far>20</far>
-        </clip>
-      </camera>
-      <ignition_frame_id>rgb_camera_link_optical_frame</ignition_frame_id>
-      <always_on>1</always_on>
-      <update_rate>30</update_rate>
-      <visualize>true</visualize>
-      <topic>camera</topic>
-    </sensor>
-  </gazebo>
-
-  <gazebo reference="segmentation_camera_link_optical_frame">
-    <sensor name="segmentation_camera" type="segmentation">
-      <camera>
-        <segmentation_type>instance</segmentation_type>
-        <horizontal_fov>1.047</horizontal_fov>
-        <image>
-          <width>320</width>
-          <height>240</height>
-        </image>
-        <clip>
-          <near>0.1</near>
-          <far>100</far>
-        </clip>
-      </camera>
-      <ignition_frame_id>segmentation_camera_link_optical_frame</ignition_frame_id>
-      <always_on>1</always_on>
-      <update_rate>30</update_rate>
-      <visualize>true</visualize>
-      <topic>segmentation</topic>
-    </sensor>
-  </gazebo>
-
-
-  <gazebo reference="right_caster_wheel">
-    <mu1>0.001</mu1>
-    <mu2>0.001</mu2>
-  </gazebo>
-
-  <gazebo reference="left_caster_wheel">
-    <mu1>0.001</mu1>
-    <mu2>0.001</mu2>
-  </gazebo>
-
-  <gazebo reference="right_wheel">
-    <mu1>100.0</mu1>
-    <mu2>100.0</mu2>
-  </gazebo>
-
-  <gazebo reference="left_wheel">
-    <mu1>100.0</mu1>
-    <mu2>100.0</mu2>
-  </gazebo>
-
-
-</robot>
