--- rae_gazebo/models/RAE/model.config
@@ -1,9 +1,0 @@
-<?xml version='1.0'?>
-<model>
-  <name>RAE</name>
-  <version>1.0</version>
-  <sdf version='1.9'>model.sdf</sdf>
-  <description>
-       WIP Fuel model
-  </description>
-</model>
