--- rae_gazebo/models/RAE/model.sdf
@@ -1,217 +1,0 @@
-<?xml version='1.0' ?>
-<sdf version='1.9'>
-    <model name='rae'>
-      <link name='base_link'>
-       <sensor name="imu_sensor" type="imu">
-    		<always_on>1</always_on>
-   		 <update_rate>1</update_rate>
-    		<visualize>true</visualize>
-    		<topic>imu</topic>
-	</sensor>
-        <inertial>
-          <pose>0.044959 0 0.009 0 -0 0</pose>
-          <mass>0.29999999999999999</mass>
-          <inertia>
-            <ixx>0.000487508</ixx>
-            <ixy>2.0208100000000002e-09</ixy>
-            <ixz>4.2612199999999998e-05</ixz>
-            <iyy>0.00041417599999999999</iyy>
-            <iyz>1.4956499999999999e-10</iyz>
-            <izz>0.00084455000000000001</izz>
-          </inertia>
-        </inertial>
-        <collision name='base_link_fixed_joint_lump__chassis_collision'>
-          <pose>0.0307 0 0 1.57 -5.551115123125783e-17 1.57</pose>
-          <geometry>
-            <mesh>
-              <scale>1 1 1</scale>
-              <uri>model://RAE/meshes/RAE.stl</uri>
-            </mesh>
-          </geometry>
-        </collision>
-        <collision name='base_link_fixed_joint_lump__left_caster_wheel_collision_1'>
-          <pose>0.0735 0.0445 -0.001 0 0 0</pose>
-          <geometry>
-            <mesh>
-              <scale>1 1 1</scale>
-              <uri>model://RAE/meshes/caster-wheel.stl</uri>
-            </mesh>
-          </geometry>
-        </collision>
-        <collision name='base_link_fixed_joint_lump__right_caster_wheel_collision_2'>
-          <pose>0.0735 -0.0445 -0.001 0 0 0</pose>
-          <geometry>
-            <mesh>
-              <scale>1 1 1</scale>
-              <uri>model://RAE/meshes/caster-wheel.stl</uri>
-            </mesh>
-          </geometry>
-        </collision>
-        <visual name='base_link_fixed_joint_lump__chassis_visual'>
-          <pose>0.0307 0 0 1.57 -5.551115123125783e-17 1.57</pose>
-          <geometry>
-            <mesh>
-              <scale>1 1 1</scale>
-              <uri>model://RAE/meshes/RAE.stl</uri>
-            </mesh>
-          </geometry>
-          <material>
-            <diffuse>0.125 0.125 0.125 1</diffuse>
-            <ambient>0.125 0.125 0.125 1</ambient>
-          </material>
-        </visual>
-        <visual name='base_link_fixed_joint_lump__left_caster_wheel_visual_1'>
-          <pose>0.0735 0.0445 -0.001 0 0 0</pose>
-          <geometry>
-            <mesh>
-              <scale>1 1 1</scale>
-              <uri>model://RAE/meshes/caster-wheel.stl</uri>
-            </mesh>
-          </geometry>
-          <material>
-            <diffuse>0.125 0.125 0.125 1</diffuse>
-            <ambient>0.125 0.125 0.125 1</ambient>
-          </material>
-        </visual>
-        <visual name='base_link_fixed_joint_lump__right_caster_wheel_visual_2'>
-          <pose>0.0735 -0.0445 -0.001 0 0 0</pose>
-          <geometry>
-            <mesh>
-              <scale>1 1 1</scale>
-              <uri>model://RAE/meshes/caster-wheel.stl</uri>
-            </mesh>
-          </geometry>
-          <material>
-            <diffuse>0.125 0.125 0.125 1</diffuse>
-            <ambient>0.125 0.125 0.125 1</ambient>
-          </material>
-        </visual>
-        <pose>0 0 0 0 -0 0</pose>
-        <enable_wind>false</enable_wind>
-      </link>
-      <joint name='left_wheel_joint' type='revolute'>
-        <pose>0 0 0 0 -0 0</pose>
-        <parent>base_link</parent>
-        <child>left_wheel</child>
-        <axis>
-          <xyz>-1 0 0</xyz>
-          <limit>
-            <lower>-10000000000000000</lower>
-            <upper>10000000000000000</upper>
-            <effort>inf</effort>
-            <velocity>inf</velocity>
-            <stiffness>100000000</stiffness>
-            <dissipation>1</dissipation>
-          </limit>
-          <dynamics>
-            <spring_reference>0</spring_reference>
-            <spring_stiffness>0</spring_stiffness>
-            <damping>0</damping>
-            <friction>0</friction>
-          </dynamics>
-        </axis>
-      </joint>
-      <link name='left_wheel'>
-        <pose>0 0.0427 0.0052 -0 -0 -1.57</pose>
-        <inertial>
-          <pose>-0.005 0 0 0 -0 0</pose>
-          <mass>0.050000000000000003</mass>
-          <inertia>
-            <ixx>3.9062500000000001e-06</ixx>
-            <ixy>0</ixy>
-            <ixz>0</ixz>
-            <iyy>2.3697899999999999e-06</iyy>
-            <iyz>0</iyz>
-            <izz>2.3697899999999999e-06</izz>
-          </inertia>
-        </inertial>
-        <collision name='left_wheel_collision'>
-          <pose>-0.0125 0 0 0 0 0</pose>
-          <geometry>
-            <mesh>
-              <scale>1 1 1</scale>
-              <uri>model://RAE/meshes/Motor-wheel.stl</uri>
-            </mesh>
-          </geometry>
-        </collision>
-        <visual name='left_wheel_visual'>
-          <pose>-0.0125 0 0 0 0 0</pose>
-          <geometry>
-            <mesh>
-              <scale>1 1 1</scale>
-              <uri>model://RAE/meshes/Motor-wheel.stl</uri>
-            </mesh>
-          </geometry>
-        </visual>
-        <enable_wind>false</enable_wind>
-      </link>
-      <joint name='right_wheel_joint' type='revolute'>
-        <pose>0 -0 0 0 -0 0</pose>
-        <parent>base_link</parent>
-        <child>right_wheel</child>
-        <axis>
-          <xyz>1 0 0</xyz>
-          <limit>
-            <lower>-10000000000000000</lower>
-            <upper>10000000000000000</upper>
-            <effort>inf</effort>
-            <velocity>inf</velocity>
-            <stiffness>100000000</stiffness>
-            <dissipation>1</dissipation>
-          </limit>
-          <dynamics>
-            <spring_reference>0</spring_reference>
-            <spring_stiffness>0</spring_stiffness>
-            <damping>0</damping>
-            <friction>0</friction>
-          </dynamics>
-        </axis>
-      </joint>
-      <link name='right_wheel'>
-        <pose>0 -0.0427 0.0052 -0 0 1.57</pose>
-        <inertial>
-          <pose>-0.005 0 0 0 -0 0</pose>
-          <mass>0.050000000000000003</mass>
-          <inertia>
-            <ixx>3.9062500000000001e-06</ixx>
-            <ixy>0</ixy>
-            <ixz>0</ixz>
-            <iyy>2.3697899999999999e-06</iyy>
-            <iyz>0</iyz>
-            <izz>2.3697899999999999e-06</izz>
-          </inertia>
-        </inertial>
-        <collision name='right_wheel_collision'>
-          <pose>-0.0125 0 0 0 0 0</pose>
-          <geometry>
-            <mesh>
-              <scale>1 1 1</scale>
-              <uri>model://RAE/meshes/Motor-wheel.stl</uri>
-            </mesh>
-          </geometry>
-        </collision>
-        <visual name='right_wheel_visual'>
-          <pose>-0.0125 0 0 0 0 0</pose>
-          <geometry>
-            <mesh>
-              <scale>1 1 1</scale>
-              <uri>model://RAE/meshes/Motor-wheel.stl</uri>
-            </mesh>
-          </geometry>
-        </visual>
-        <enable_wind>false</enable_wind>
-      </link>
-      <static>false</static>
-      <plugin name='ignition::gazebo::systems::DiffDrive' filename='libignition-gazebo-diff-drive-system.so'>
-        <left_joint>left_wheel_joint</left_joint>
-        <right_joint>right_wheel_joint</right_joint>
-        <wheel_separation>0.0854</wheel_separation>
-        <wheel_radius>0.01</wheel_radius>
-        <odom_publish_frequency>1</odom_publish_frequency>
-        <topic>cmd_vel</topic>
-      </plugin>
-      <pose>0 0 0.005887 -0 0 -0</pose>
-      <self_collide>false</self_collide>
-    </model>
-    
-</sdf>
