--- rae_hw/include/rae_hw/rae_hw.hpp
@@ -1,84 +1,0 @@
-// Copyright (c) 2023, Luxonis
-// Copyright (c) 2023, Stogl Robotics Consulting UG (haftungsbeschränkt) (template)
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef RAE_HW__RAE_HW_HPP_
-#define RAE_HW__RAE_HW_HPP_
-
-#include <string>
-#include <vector>
-#include <chrono>
-
-#include "rae_hw/rae_motors.hpp"
-#include "rae_hw/visibility_control.h"
-#include "hardware_interface/system_interface.hpp"
-#include "hardware_interface/handle.hpp"
-#include "hardware_interface/hardware_info.hpp"
-#include "hardware_interface/types/hardware_interface_return_values.hpp"
-#include "rclcpp/macros.hpp"
-#include "rclcpp/rclcpp.hpp"
-#include "rclcpp_lifecycle/state.hpp"
-#include "gpiod.hpp"
-
-namespace rae_hw
-{
-  class RaeHW : public hardware_interface::SystemInterface
-  {
-  public:
-    TEMPLATES__ROS2_CONTROL__VISIBILITY_PUBLIC
-    hardware_interface::CallbackReturn on_init(
-        const hardware_interface::HardwareInfo &info) override;
-
-    TEMPLATES__ROS2_CONTROL__VISIBILITY_PUBLIC
-    hardware_interface::CallbackReturn on_configure(
-        const rclcpp_lifecycle::State &previous_state) override;
-
-    TEMPLATES__ROS2_CONTROL__VISIBILITY_PUBLIC
-    std::vector<hardware_interface::StateInterface> export_state_interfaces() override;
-
-    TEMPLATES__ROS2_CONTROL__VISIBILITY_PUBLIC
-    std::vector<hardware_interface::CommandInterface> export_command_interfaces() override;
-
-    TEMPLATES__ROS2_CONTROL__VISIBILITY_PUBLIC
-    hardware_interface::CallbackReturn on_activate(
-        const rclcpp_lifecycle::State &previous_state) override;
-
-    TEMPLATES__ROS2_CONTROL__VISIBILITY_PUBLIC
-    hardware_interface::CallbackReturn on_deactivate(
-        const rclcpp_lifecycle::State &previous_state) override;
-
-    TEMPLATES__ROS2_CONTROL__VISIBILITY_PUBLIC
-    hardware_interface::CallbackReturn on_shutdown(
-        const rclcpp_lifecycle::State &previous_state) override;
-
-    TEMPLATES__ROS2_CONTROL__VISIBILITY_PUBLIC
-    hardware_interface::return_type read(
-        const rclcpp::Time &time, const rclcpp::Duration &period) override;
-
-    TEMPLATES__ROS2_CONTROL__VISIBILITY_PUBLIC
-    hardware_interface::return_type write(
-        const rclcpp::Time &time, const rclcpp::Duration &period) override;
-
-  private:
-    std::unique_ptr<RaeMotor> motorL, motorR;
-    double leftMotorCMD, rightMotorCMD;
-    int pwmA, pwmB, phA, phB;
-    double leftPos, rightPos, leftVel, rightVel;
-    std::string leftWheelName, rightWheelName;
-    std::chrono::high_resolution_clock::time_point prevTime;
-  };
-
-} // namespace rae_hw
-
-#endif // RAE_HW__RAE_HW_HPP_
