--- rae_hw/launch/rae_control.launch.py
@@ -1,74 +1,0 @@
-# Copyright 2020 ROS2-Control Development Team (2020)
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import os
-
-from ament_index_python.packages import get_package_share_directory
-
-from launch import LaunchDescription
-from launch.actions import IncludeLaunchDescription
-from launch.launch_description_sources import PythonLaunchDescriptionSource
-from launch_ros.actions import Node
-
-import xacro
-
-
-def generate_launch_description():
-
-    # Get URDF via xacro
-    robot_description_path = os.path.join(
-        get_package_share_directory('rae_description'),
-        'urdf',
-        'rae.urdf.xacro')
-    robot_description_config = xacro.process_file(robot_description_path)
-    robot_description = {'robot_description': robot_description_config.toxml()}
-
-    test_controller = os.path.join(
-        get_package_share_directory('rae_hw'),
-        'config',
-        'controller.yaml'
-    )
-
-    return LaunchDescription([
-        IncludeLaunchDescription(
-            PythonLaunchDescriptionSource(
-                os.path.join(get_package_share_directory('rae_description'), 'launch', 'rae_gazebo_desc_launch.py')),
-            launch_arguments={"sim": "false"}.items()
-        ),
-        Node(
-            package='controller_manager',
-            executable='ros2_control_node',
-            parameters=[robot_description, test_controller],
-            remappings=[('/diff_controller/cmd_vel_unstamped', 'cmd_vel')],
-            output={
-                'stdout': 'screen',
-                'stderr': 'screen',
-            },
-        ),
-        Node(
-            package="controller_manager",
-            executable="spawner",
-            arguments=["diff_controller", "-c", "/controller_manager"],
-        ),
-        Node(
-            package="controller_manager",
-            executable="spawner",
-            arguments=["joint_state_broadcaster",
-                       "--controller-manager", "/controller_manager"],
-        ),
-        Node(
-            package="rae_hw",
-            executable="battery",
-        )
-    ])
