--- rae_hw/package.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>rae_hw</name>
-  <version>0.0.0</version>
-  <description>Package containing ROS2 Hardware Interface for RAE</description>
-  <maintainer email="adam.serafin@luxonis.com">Adam Serafin</maintainer>
-  <license>MIT</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <depend>hardware_interface</depend>
-
-  <depend>libgpiod-dev</depend>
-
-  <depend>pluginlib</depend>
-
-  <depend>rclcpp</depend>
-
-  <depend>rclcpp_lifecycle</depend>
-
-  <depend>controller_manager</depend>
-  <depend>diff_drive_controller</depend>
-  <depend>ros2_controllers</depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-  <test_depend>ament_cmake_gmock</test_depend>
-  <test_depend>ros2_control_test_assets</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
