--- rae_hw/rae_hw.xml
@@ -1,9 +1,0 @@
-<library path="rae_hw">
-  <class name="rae_hw/RaeHW"
-         type="rae_hw::RaeHW"
-         base_class_type="hardware_interface::SystemInterface">
-    <description>
-      ros2_control hardware interface.
-    </description>
-  </class>
-</library>
